Small fixes after eval

+++ Module09/ex00/src/BitcoinExchange.cpp
@@ -6,7 +6,7 @@
 /*   By: fkoolhov <fkoolhov@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/22 15:25:42 by fkoolhov          #+#    #+#             */
-/*   Updated: 2024/03/06 14:01:20 by fkoolhov         ###   ########.fr       */
+/*   Updated: 2024/03/06 18:49:00 by fkoolhov         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,7 +26,7 @@ BitcoinExchange::BitcoinExchange(std::ifstream& datafile)
 		date_stream >> std::get_time(&date_struct, "%Y-%m-%d");
 		int date_as_int = (date_struct.tm_year + 1900) * 10000 + (date_struct.tm_mon + 1) * 100 + date_struct.tm_mday;
 		std::string rate_string = line.substr(comma_position + 1);
-		double rate = std::stof(rate_string);
+		double rate = std::stod(rate_string);
 		database.emplace(date_as_int, rate);
 	}
 
@@ -112,7 +112,7 @@ bool BitcoinExchange::InputIsValid(const std::string& line)
 		if (VerifyDate(date_struct, line))
 		{
 			this->date_as_int = (date_struct.tm_year + 1900) * 10000 + (date_struct.tm_mon + 1) * 100 + date_struct.tm_mday;
-			this->amount = std::stof(amount_string);
+			this->amount = std::stod(amount_string);
 			
 			if (this->amount < 0)
 				std::cerr << "Error: not a positive number.\n";
